+++ src/main/kotlin/com/dsoftware/ghmanager/ui/panels/wfruns/filters/WfRunsFiltersFactory.kt
@@ -1,4 +1,4 @@
-package com.dsoftware.ghmanager.ui.panels.filters
+package com.dsoftware.ghmanager.ui.panels.wfruns.filters
 
 import com.dsoftware.ghmanager.ui.ToolbarUtil
 import com.intellij.collaboration.ui.codereview.list.search.ChooserPopupUtil
