+++ src/main/kotlin/com/dsoftware/ghmanager/ui/panels/wfruns/filters/WfRunsSearchHistoryModel.kt
@@ -1,4 +1,4 @@
-package com.dsoftware.ghmanager.ui.panels.filters
+package com.dsoftware.ghmanager.ui.panels.wfruns.filters
 
 import com.intellij.collaboration.ui.codereview.list.search.PersistingReviewListSearchHistoryModel
